+++ packages/BiostructureViewer/src/package-test.ts
@@ -8,6 +8,7 @@ import './tests/pdb-helper-tests';
 import './tests/pdb-tests';
 import './tests/pdb-grid-cell-renderer-tests';
 import './tests/molstar-preview-tests';
+import './tests/molstar-viewer-tests';
 import './tests/viewers';
 import './tests/pdbqt-tests';