+++ packages/BiostructureViewer/src/tests/viewers.ts
@@ -5,7 +5,7 @@ import * as ui from 'datagrok-api/ui';
 import {category, test, testViewer} from '@datagrok-libraries/utils/src/test';
 import {IPdbHelper} from '@datagrok-libraries/bio/src/pdb/pdb-helper';
 
-import {_packageName} from './utils';
+import {_package} from '../package-test';
 
 category('Viewers', () => {
   // -- Viewers tests --
@@ -14,7 +14,7 @@ category('Viewers', () => {
     test(v, async () => {
       await testViewer(v, await (async () => {
         const ph: IPdbHelper = await grok.functions.call('BiostructureViewer:getPdbHelper');
-        const pdbFn: string = `System:AppData/${_packageName}/samples/1bdq.pdb`;
+        const pdbFn: string = `System:AppData/${_package.name}/samples/1bdq.pdb`;
         const pdbStr: string = await grok.dapi.files.readAsText(pdbFn);
         const df: DG.DataFrame = await ph.pdbToDf(pdbStr, '1bdq');
         return df;