+++ packages/BiostructureViewer/src/utils/pdbqt/import-pdbqt.ts
@@ -8,12 +8,12 @@ import {TAGS as pdbTAGS} from '@datagrok-libraries/bio/src/pdb/index';
 
 import {Pdbqt} from './pdbqt-parser';
 import {IMPORT} from '../../consts-import';
-import {PROPS as bsPROPS} from '../../viewers/molstar-viewer/molstar-viewer';
+import {buildDataJson, PROPS as bsPROPS} from '../../viewers/molstar-viewer/molstar-viewer';
 import {Base64} from 'js-base64';
 import {viewBiostructure} from '../../viewers/view-preview';
 
 
-export async function importPdbqtUI(fileContent: string): Promise<DG.DataFrame[]> {
+export async function importPdbqtUI(fileContent: string, test: boolean = false): Promise<DG.DataFrame[]> {
   const data: Pdbqt = Pdbqt.parse(fileContent);
 
   if (data.models.length > 0) {
@@ -22,8 +22,10 @@ export async function importPdbqtUI(fileContent: string): Promise<DG.DataFrame[]
     if (data.target) {
       const targetPdbStr = data.target.toPdb();
       df.setTag(pdbTAGS.PDB, targetPdbStr);
-      return [];
+      return [df];
     } else {
+      if (test) return [df];
+
       const dataFileProp = DG.Property.fromOptions({name: 'dataFile', caption: 'Data file', type: 'file'});
       const dataFileInput = DG.InputBase.forProperty(dataFileProp);
 
@@ -36,7 +38,7 @@ export async function importPdbqtUI(fileContent: string): Promise<DG.DataFrame[]
           const dataFi: DG.FileInfo = dataFileInput.value;
           const dataA: Uint8Array = dataFi.data ? dataFi.data /* User's file*/ : await dataFi.readAsBytes()/* Shares */;
           const viewer = await df.plot.fromType('Biostructure', {
-            [bsPROPS.dataJson]: JSON.stringify({data: Base64.fromUint8Array(dataA), ext: dataFi.extension}),
+            [bsPROPS.dataJson]: buildDataJson(dataA, dataFi.extension),
             [bsPROPS.ligandColumnName]: molColName,
           });
           view.dockManager.dock(viewer, DG.DOCK_TYPE.RIGHT, null, 'Docking target biostructure', 0.4);