+++ packages/BiostructureViewer/CHANGELOG.md
@@ -15,6 +15,8 @@
 ### Bug fixes
 
 * Fix caching for PdbGridCellRenderer
+* Fix MolstarViewer for Molecule ligands, tests
+* Fix pdbGridCellRenderer, test
 
 ## 1.0.11 (2023-07-24)